+++ example/app.py
@@ -20,7 +20,7 @@ security = [
 ]
 
 authorizations = {
-    "apikey": {
+    "api_key": {
         "type": "apiKey",
         "in": "header",
         "name": "Authorization",
@@ -30,7 +30,7 @@ authorizations = {
 }
 
 servers = [{"url": "http://localhost:5001"}]
-api = Api(app, version='5', servers=servers, title="APP")
+api = Api(app, version='5', servers=servers, title="APP", authorizations=authorizations)
 
 
 def auth(api_key, endpoint, method):
@@ -47,8 +47,7 @@ API_URL = 'swagger.json'  # Our API url (can of course be a local resource)
 swagger_blueprint = get_swagger_blueprint(
     api.open_api_object,
     swagger_prefix_url=SWAGGER_URL,
-    swagger_url=API_URL,
-    authorizations=authorizations)
+    swagger_url=API_URL)
 
 
 swagger.auth = auth